requirements.txt
discord.py==2.0.1
packaging==21.3
python-dotenv==0.20.0
Pillow==9.4.0
banana-dev==4.0.1